chore: delete the recommendations app, move its FT layer to utils
Follow-up to `317f1972b` (PR 8695), which removed the cancelled SPS
recommendations beta but had to keep the app package, its migrations and its
INSTALLED_APPS entry until `recommendations.0002_delete_beneficiary` had run
everywhere.
