--- itou/recommendations/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class RecommendationsConfig(AppConfig):
-    name = "itou.recommendations"
-    verbose_name = "Recommandations"
